+++ internal/store/queue.go
@@ -3,8 +3,8 @@ package store
 import "time"
 
 type Queue struct {
-	ID         int64     `json:"id"`
-	WorkflowID int64     `json:"workflow_id"`
+	ID         int     `json:"id"`
+	WorkflowID int     `json:"workflow_id"`
 	Status     string    `json:"status"`
 	EnqueuedAt time.Time `json:"enqueued_at"`
 }
